+++ src/LIN3S/SharedKernel/Domain/Model/DateTime/DateTime.php
@@ -13,38 +13,9 @@ declare(strict_types=1);
 
 namespace LIN3S\SharedKernel\Domain\Model\DateTime;
 
-final class DateTime
+/**
+ * @author Beñat Espiña <benatespina@gmail.com>
+ */
+final class DateTime extends \DateTimeImmutable
 {
-    public static function now()
-    {
-        return new \DateTimeImmutable();
-    }
-
-    public static function fromTime($time)
-    {
-        return new \DateTimeImmutable($time);
-    }
-
-    public static function fromRelative($relative)
-    {
-        return new \DateTimeImmutable($relative);
-    }
-
-    public static function utcTimestamp()
-    {
-        return (new \DateTimeImmutable('now', new \DateTimeZone('UTC')))->getTimestamp();
-    }
-
-    public static function timestamp()
-    {
-        return (new \DateTimeImmutable())->getTimestamp();
-    }
-
-    private function __construct()
-    {
-    }
-
-    private function __clone()
-    {
-    }
 }