+++ src/LIN3S/SharedKernel/Infrastructure/Persistence/Sql/Pdo.php
@@ -25,36 +25,45 @@ final class Pdo
     public function __construct(\PDO $pdo)
     {
         $this->pdo = $pdo;
+        $this->pdo->setAttribute($this->pdo::ATTR_ERRMODE, $this->pdo::ERRMODE_EXCEPTION);
     }
 
-    public function exec($statement) : int
+    public function query(string $sql, array $parameters) : array
     {
-        return $this->pdo->exec($statement);
+        return $this->execute($sql, $parameters)->fetchAll(\PDO::FETCH_ASSOC);
     }
 
-    public function beginTransaction() : void
+    public function insert($table, $columns, $numberOfInsertions, callable $prepareData) : void
+    {
+        $rowPlaces = '(' . implode(', ', array_fill(0, count($columns), '?')) . ')';
+        $allPlaces = implode(', ', array_fill(0, $numberOfInsertions, $rowPlaces));
+
+        $sql = 'INSERT INTO ' . $table . ' (' . implode(', ', $columns) . ') VALUES ' . $allPlaces;
+
+        $this->execute($sql, call_user_func($prepareData));
+    }
+
+    public function executeAtomically(callable $function)
     {
         $this->pdo->beginTransaction();
+
+        try {
+            $return = call_user_func($function, $this);
+
+            $this->pdo->commit();
+
+            return $return ?: true;
+        } catch (\Exception | \Throwable $exception) {
+            $this->pdo->rollback();
+
+            throw $exception;
+        }
     }
 
-    public function commit() : void
-    {
-        $this->pdo->commit();
-    }
-
-    public function rollback() : void
-    {
-        $this->pdo->rollBack();
-    }
-
-    public function execute($sql, array $parameters) : \PDOStatement
+    public function execute(string $sql, array $parameters) : \PDOStatement
     {
         $statement = $this->pdo->prepare($sql);
-        $result = $statement->execute($parameters);
-
-        if (false === $result) {
-            throw new PdoExecutionFailed($statement->errorInfo());
-        }
+        $statement->execute($parameters);
 
         return $statement;
     }